dashboard/pyproject.toml
[project]
name = "iv127-spring-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anywidget>=0.9.18",
    "bertopic==0.17.0",
    "corpy>=0.6.1",
    "datamapplot>=0.5.1",
    "faicons>=0.2.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "itables>=2.2.5",
    "jupyter>=1.1.1",
    "litellm>=1.65.1",
    "llvmlite==0.44.0",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "plotly==5.24",
    "plotnine>=0.14.5",
    "polars>=1.23.0",
    "quarto>=0.1.0",
    "ridgeplot>=0.3.2",
    "shiny>=1.3.0",
    "shinywidgets>=0.5.1",
    "ufal-morphodita>=1.11.2.1",
    "umap-learn>=0.5.7",
    "umap>=0.1.1",
]
